LPS-72641 fixed selected attribute on option tag
According to the MDN documentation:
https://developer.mozilla.org/en/docs/Web/HTML/Element/option
"If present, this Boolean attribute indicates that the option is
initially selected"

This means that this attribute should not have a value of true or false,
but just be present or not. I have tested this behaviour on Chrome,
Safari and Firefox and it seems to work.