LPS-139522 Providing a way to reindex an item without notifying the contributors.
JournalArticle versions are persisted in the same table, so they are the same base model. When an article is indexed we should reindex the other versions of the same article to keep updated the indexed documents. For example, when a new article version is published, depending on the configured value in "Index All Article Versions Enabled, the previous indexed version:
- If disabled: must be deleted.
- If enabled: must be updated because when publishing a new version the previous version will not be HEAD anymore.

If we don't provide a way to reindex without notifying the contributor, when we try to reindex a journal article from the contributor callback method will cause an infinity loop. That is because we can't determine when an article is reindexing because it has changed or because another version of the same article has changed.