LPS-156173 Add a method to determine if the form contains a captcha field.
We are checking only the items contained by the form item for better performance and to prevent the case that it could be more than one form on the page.

It is implemented by iterating recursively for the items contained by the form item to check if its FragmentEntry has enabled the captcha typeOptions.